Add files via upload
In today's highly competitive production industry, customer satisfaction is a major concern for service providers. Customer churn occurs when customers switch to a competitor's service, resulting in lost revenue for the provider. In this project, for predicting milk quality and improving the quality and safety of dairy products. we aim to develop a machine learning model that can predict customer churn, thereby enabling the provider to take proactive measures to retain customers.
We will use a dataset that contains milk ingredients information such as PH, turbidity, colour, odour. We will preprocess the data by cleaning it and performing feature engineering to extract meaningful information from the raw data. We will then split the data 


into training and testing sets to train and evaluate our machine learning model.
We will explore different classification algorithms such as Naïve bayes, K-Nearest neighbour, decision trees, and support vector machines to determine which algorithm provides the best accuracy in predicting customer churn.
Finally, we will evaluate the performance of our model using metrics such as accuracy, precision, recall, and F1-score. We can also perform a cost-benefit analysis to estimate the financial impact of our model on the provider's business.
Overall, this project aims to demonstrate the effectiveness of machine learning in predicting customer churn in the service sector industry and to provide insights to the provider for customer retention strategies.